feat: Improve handling of selected option in main button click event
This commit improves the handling of the selected option in the main button click event in the Telegram WebApp. It adds console logs to indicate the start of the event and the selected option. If the selected option is included in the predefined reasons, it is assigned to the 'answer' variable. Otherwise, the 'otherReason' value is assigned. The 'answer' value is then used to construct a transmitting string that is sent back to the bot. This change enhances the functionality and ensures the correct data is transmitted.